fix types

+++ tasks/seq/Sadikov_I_SparesMatrixMultiplication/src/SparesMatrix.cpp
@@ -12,9 +12,9 @@ SparesMatrix SparesMatrix::Transpose(const SparesMatrix& matrix) {
   std::vector<std::vector<double>> intermediate_values(max_size);
   std::vector<std::vector<int>> intermediate_indexes(max_size);
   int counter = 0;
-  for (size_t i = 0; i < matrix.GetElementsSum().size(); ++i) {
+  for (int i = 0; i < static_cast<int>(matrix.GetElementsSum().size()); ++i) {
     auto limit = i == 0 ? matrix.GetElementsSum()[0] : matrix.GetElementsSum()[i] - matrix.GetElementsSum()[i - 1];
-    for (size_t j = 0; j < limit; ++j) {
+    for (int j = 0; j < limit; ++j) {
       intermediate_values[matrix.GetRows()[counter]].emplace_back(matrix.GetValues()[counter]);
       intermediate_indexes[matrix.GetRows()[counter]].emplace_back(i);
       counter++;
